HexFiend: silence -Wimplicit-const-int-float-conversion in HFFPToUL
clang 21 (Xcode 26.5) treats -Wimplicit-const-int-float-conversion as
an error by default. The assert at HFFunctions.h:357 was

    assert(val <= ULLONG_MAX);

where `val` is `long double` and `ULLONG_MAX` is an unsigned long long
constant — the implicit conversion rounds 18446744073709551615 to
18446744073709551616 (long double can't represent ULLONG_MAX exactly),
which the warning flags.

Make the conversion explicit:

    assert(val <= (long double)ULLONG_MAX);

This silences the warning without changing observable behavior; the
assert remains a debugging safety net for inputs near the
unsigned-long-long ceiling, and the actual conversion on the next line
is unchanged. Surfaced while building PR #3 on a Mac with Xcode 26.5;
the same fix unblocks any Cocoa build on that toolchain.

+++ HexFiend/HFFunctions.h
@@ -354,7 +354,7 @@ static inline NSUInteger HFAtomicDecrement(volatile NSUInteger *ptr, BOOL barrie
 /*! Converts a long double to unsigned long long.  Assumes that val is already an integer - use floorl or ceill */
 static inline unsigned long long HFFPToUL(long double val) {
     assert(val >= 0);
-    assert(val <= ULLONG_MAX);
+    assert(val <= (long double)ULLONG_MAX);
     unsigned long long result = (unsigned long long)val;
     assert((long double)result == val);
     return result;
